Add in abstract base class to reduce code duplication
